added more folders and move Navbar to views
