+++ client/src/pages/MainPage.jsx
@@ -1,6 +1,6 @@
 import React from "react";
 import { Outlet } from "react-router-dom";
-import Navbar from "../Navbar/Navbar";
+import Navbar from "../views/Navbar";
 import "./MainPage.scss";
 
 const MainPage = () => {
